+++ qupath-gui-fx/src/main/java/qupath/lib/gui/commands/TMAGridView.java
@@ -25,6 +25,7 @@ package qupath.lib.gui.commands;
 
 import java.awt.image.BufferedImage;
 import java.text.DecimalFormat;
+import java.text.NumberFormat;
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.LinkedHashMap;
@@ -506,7 +507,7 @@ public class TMAGridView implements PathCommand, ImageDataChangeListener<Buffere
 			
 			if (measurement.getValue() != null && showMeasurement.getValue()) {
 				double value = model.getNumericValue(core, measurement.getValue());
-				setText(GeneralTools.createFormatter(2).format(value));
+				setText(GeneralTools.formatNumber(value, 2));
 				setTextAlignment(TextAlignment.CENTER);
 				setAlignment(Pos.BOTTOM_CENTER);
 			} else
@@ -595,13 +596,12 @@ public class TMAGridView implements PathCommand, ImageDataChangeListener<Buffere
 		
 		void updateMeasurementText() {
 			String m = measurement == null ? null : measurement.get();
-			DecimalFormat formatter = GeneralTools.createFormatter(3);
 			for (Entry<TMACoreObject, Label> entry : nodeMap.entrySet()) {
 				if (m == null || !showMeasurement.get())
 					entry.getValue().setText(" ");
 				else {
 					double val = model.getNumericValue(entry.getKey(), m);
-					entry.getValue().setText(formatter.format(val));
+					entry.getValue().setText(GeneralTools.formatNumber(val, 3));
 				}
 				entry.getValue().setContentDisplay(ContentDisplay.TOP);
 			}